Debug

+++ CPU.h
@@ -12,7 +12,7 @@ using namespace std;
 class CPU{
 	public:
 		uint16 pc; // program counter
-		uint8 sp;  // stack pointer
+		uint16 sp;  // stack pointer
 		uint8 A;   // accumulator
 		uint8 X;  // index register X
 		uint8 Y;  // index register Y
